backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
python-jose==3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv==1.0.1
email-validator==2.1.0
bcrypt==4.1.3
